Refactor Rendering of Objects (prepare for CPU rendering)
New abstract class ObjectRenderer and derived classes for each type of Object.
It's a bit more complex but hopefully easier to build on now.

There are probably a heap of bugs in this, but I can see the test pattern again, so I'll commit before it gets worse.

Note: We now have to make sure Screen is initialised first or the segfaults will hit the fan.
	(Now it makes sense why all those things weren't in constructors in the first place :S)

It also now segfaults if you get View and Screen the wrong way round.