Go nuts with Qt
We can do all the things I promised!*
And we can do it without having to implement a vim style stdio based interface!
Or adding 16 extra mouse buttons!

Qt can parse XML or even SVG all by itself though...
I'm going to ignore that and just keep treating it as a menu system.

* Commit --amend

Well, the set of things we can do is neither a subset nor a superset of the things I promised.