Add MinGW Win32 Cross Compiling
Not tested with Qt4 yet, but works with QUADTREE_DISABLED

Build procedure:
 1. Put source for GMP and SDL2 in ipdf/code/contrib/
    Do 2. and 3. for both
 2. Run `./configure --host=i686-w64-mingw32 --prefix="ipdf/code/contrib/win32"`
 3. `make; make install`
 4. In ipdf/code/src, run `make ARCH=win32 CONTROLPANEL=disabled`
 5. Download SDL2.dll for the binary

Will try and get Qt4 support in the morning I guess.

I am doing this because I want it to be easy for marker(s) to run the software.
So I will make a precompiled ipdf.exe for windows in addition to the linux binary.

Tested under wine, seems to work.

ALSO

Why the hell is there a goto in document.cpp